fixed compilation with SIMD instructions fixed timer1 tutorial project
git-svn-id: https://svn.grrrr.org/ext/trunk@619 4d9ac71a-51e6-0310-8455-cad1006bcd31

+++ tutorial/timer1/main.cpp
@@ -39,8 +39,8 @@ protected:
 	// timers
 	Timer tmrA,tmrB;
 
-	void m_getostime(float &f) { f = GetOSTime(); }  // method for operating system time attribute
-	void m_getrttime(float &f) { f = GetTime(); }  // method for real-time system time attribute
+	void m_getostime(float &f) { f = (float)GetOSTime(); }  // method for operating system time attribute
+	void m_getrttime(float &f) { f = (float)GetTime(); }  // method for real-time system time attribute
 
 	void m_timerA(void *) { ToOutString(0,"Timer A"); }  // timer A method
 	void m_timerB(void *) { ToOutString(0,"Timer B"); }  // timer B method
